Lint
git-svn-id: https://svn.code.sf.net/p/tcnopen/trdp/trunk/trdp@1647 3b5a3598-5f4e-4449-9e63-bd40438bfec0

+++ src/vos/windows/vos_sock.c
@@ -1549,7 +1549,7 @@ EXT_DECL VOS_ERR_T vos_sockReceiveTCP (
     UINT32  *pSize)
 {
     int rcvSize     = 0;
-    int bufferSize  = (size_t) *pSize;
+    int bufferSize  = (int) *pSize;
     int err         = 0;
 
     *pSize = 0;
